+++ src/components/pagination/VueScreenerPagination.vue
@@ -39,7 +39,7 @@
 </template>
 
 <script lang="ts" setup>
-import { VueScreener } from '@/interfaces/screener'
+import { VueScreener } from '@/interfaces/vue-screener'
 import { computed, onMounted, watch } from 'vue'
 import UiButton, { ButtonUI } from './ui/button/Button.vue'
 import UiInput, { InputUI } from './ui/input/Input.vue'
