+++ src/interfaces/vue-screener.ts
